# optimizes non-windows stat to avoid opening a file