--- a/internal/platform/open_file_js.go
+++ b/internal/platform/open_file_js.go
@@ -1,5 +1,3 @@
-//go:build js
-
 package platform
 
 import (