--- a/internal/platform/stat_bsd.go
+++ b/internal/platform/stat_bsd.go
@@ -7,7 +7,20 @@ import (
 	"syscall"
 )
 
+func stat(path string, st *Stat_t) (err error) {
+	t, err := os.Stat(path)
+	if err = UnwrapOSError(err); err == nil {
+		fillStatFromSys(st, t)
+	}
+	return
+}
+
 func fillStatFromOpenFile(stat *Stat_t, fd uintptr, t os.FileInfo) (err error) {
+	fillStatFromSys(stat, t)
+	return
+}
+
+func fillStatFromSys(stat *Stat_t, t os.FileInfo) {
 	d := t.Sys().(*syscall.Stat_t)
 	stat.Ino = d.Ino
 	stat.Dev = uint64(d.Dev)
@@ -20,5 +33,4 @@ func fillStatFromOpenFile(stat *Stat_t, fd uintptr, t os.FileInfo) (err error) {
 	stat.Mtim = mtime.Sec*1e9 + mtime.Nsec
 	ctime := d.Ctimespec
 	stat.Ctim = ctime.Sec*1e9 + ctime.Nsec
-	return
 }