--- a/internal/platform/stat_linux.go
+++ b/internal/platform/stat_linux.go
@@ -10,7 +10,20 @@ import (
 	"syscall"
 )
 
+func stat(path string, st *Stat_t) (err error) {
+	t, err := os.Stat(path)
+	if err = UnwrapOSError(err); err == nil {
+		fillStatFromSys(st, t)
+	}
+	return
+}
+
 func fillStatFromOpenFile(stat *Stat_t, fd uintptr, t os.FileInfo) (err error) {
+	fillStatFromSys(stat, t)
+	return
+}
+
+func fillStatFromSys(stat *Stat_t, t os.FileInfo) {
 	d := t.Sys().(*syscall.Stat_t)
 	stat.Ino = uint64(d.Ino)
 	stat.Dev = uint64(d.Dev)
@@ -23,5 +36,4 @@ func fillStatFromOpenFile(stat *Stat_t, fd uintptr, t os.FileInfo) (err error) {
 	stat.Mtim = mtime.Sec*1e9 + mtime.Nsec
 	ctime := d.Ctim
 	stat.Ctim = ctime.Sec*1e9 + ctime.Nsec
-	return
 }