This ensures only windows needs to open a file in order to do a stat. Maybe one day it can also avoid that, but meanwhile darwin and linux should be faster.

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/imports/wasi_snapshot_preview1
                                       │   old.txt    │              new.txt               │
                                       │    sec/op    │    sec/op     vs base              │
_pathFilestat/embed.FS_fd=root-12        1.399µ ± ∞ ¹   1.388µ ± ∞ ¹       ~ (p=1.000 n=5)
_pathFilestat/embed.FS_fd=directory-12   1.448µ ± ∞ ¹   1.405µ ± ∞ ¹  -2.97% (p=0.008 n=5)
_pathFilestat/os.DirFS_fd=root-12        9.017µ ± ∞ ¹   8.886µ ± ∞ ¹       ~ (p=0.151 n=5)
_pathFilestat/os.DirFS_fd=directory-12   9.338µ ± ∞ ¹   9.375µ ± ∞ ¹       ~ (p=0.548 n=5)
geomean                                  3.614µ         3.570µ        -1.21%
```